feat(job): add getPurchaseById method to JobService and update refund request methods
